+++ CHANGELOG.md
@@ -12,11 +12,13 @@
 - `IfUnion` type
 - `swap`
 - `KeyOf` type
+- `At` type
 
 ### Changed
 - rename `isEnumerable` to `isEnumerableProp`
 - `WeakMeta` type
 - Improve `has` and `isEnumerableProp` types.
+- Improve `at` and `updateAt` types.
 
 ### Removed
 - `TypeOf` type